requirements.txt
django==1.9rc1
django_smuggler==0.6.0
pytz==2015.7
pillow==3.0.0
django-resized==0.3.5
django-ckeditor==5.0.3
